Modelling the Solar System using Newtonian Mech.
This commit contains a use case of an N-body simulation i.e. simulation of our solar system using Newtonian Dynamics. It uses the astropy package to obtain the initial parameters of the simulation, namely the position and velocity of the solar system objects and then uses an acceleration function based on Newton's law of gravitation to keep a track of solar system objects' movement. It also generates an attractive visualization of the simulation using plotly and matplotlib.